+++ priv/repo/migrations/20190903132047_private_boards.exs
@@ -1,7 +1,0 @@
-defmodule Lucidboard.Repo.Migrations.PrivateBoards do
-  use Ecto.Migration
-
-  def change do
-
-  end
-end